Move towards deprecating Config.Factory
Gitdb now embeds meta data in each record to support indexing
instead of relying on Config.Factory. This is a breaking change
as gitdb won't be able index data written by previous versions.

Also improved the encrypt/decrypt points in gitdb moving it closer
to the point of use. This way readBlock and writeBlock is simplified
and we only decrypt data when your app needs it