Move queueing system to memory
Previously gitdb maintained the write queue in a file, which meant
more complex logic when flushing to database. This has been changed
after some consideration and infact memory-based queue is simpler
to understand and works just as well. To ensure that items in the
queue are not lost in case of an app panic, developers should ensure
to call *Connection.Close() in their error handler function